+++ packages/dreki/src/entity/entities.ts
@@ -18,11 +18,11 @@ export class Entities {
   private len: number;
 
   /**
-   * Create new entity storage with given capacity. Will call `on_realloc` when reallocations occurs.
+   * Create new entity storage with given capacity. Will call `onRealloc` when reallocations occurs.
    * @param capacity
-   * @param on_realloc
+   * @param onRealloc
    */
-  constructor(capacity: number, private readonly on_realloc?: (length: number) => unknown) {
+  constructor(capacity: number, private readonly onRealloc?: (length: number) => unknown) {
     this.metadata = vec(capacity, () => ({ generation: 0 }));
     this.freelist = vec(32, 0);
     this.len = 0;
@@ -45,7 +45,7 @@ export class Entities {
         throw new Error(`Can't spawn new entity, max capacity reached (${this.capacity})`);
       }
       this.metadata.realloc(Math.min(Math.floor(this.capacity * 1.5), MAX_ENTITY_CAPACITY));
-      if (this.on_realloc) this.on_realloc(this.metadata.capacity);
+      if (this.onRealloc) this.onRealloc(this.metadata.capacity);
     }
 
     if (index === undefined) {
@@ -62,7 +62,7 @@ export class Entities {
    * @returns
    */
   public dispose(entity: Entity) {
-    const handle = Entity.handle_of(entity);
+    const handle = Entity.handleOf(entity);
     const meta = this.metadata.raw[handle.index];
 
     // don't free if there is a generation mismatch
@@ -79,7 +79,7 @@ export class Entities {
    * @returns
    */
   contains(entity: Entity): boolean {
-    const handle = Entity.handle_of(entity);
+    const handle = Entity.handleOf(entity);
     return (
       handle.index < this.len &&
       handle.index >= 0 &&
@@ -100,14 +100,14 @@ export class Entities {
    * @returns
    */
   [Symbol.iterator](): Iterator<Entity> {
-    let read_index = 0;
+    let readIndex = 0;
     const data = this.metadata;
     const length = this.len;
     return {
       next(): IteratorResult<Entity> {
-        if (read_index < length) {
+        if (readIndex < length) {
           return {
-            value: Entity(read_index, data.raw[read_index++].generation),
+            value: Entity(readIndex, data.raw[readIndex++].generation),
           };
         }
         return {