+++ packages/dreki/src/scheduler/stage.ts
@@ -3,27 +3,39 @@ import type { World } from "../mod";
 import { ResourceNotFoundError } from "../world/mod";
 import { System, SystemFunc } from "./system";
 
+/**
+ * A runnable type
+ */
 export type Runnable = {
   run(world: World): unknown;
 };
 
+/**
+ * Stage label
+ */
 export type StageLabel = string;
 
+/**
+ * A stage, have a collection of systems
+ */
 export class Stage implements Runnable {
-  public systems: System[];
+  systems: System[];
+  lastTickCheck: number;
 
-  last_tick_check: number;
-
-  public constructor(...systems: SystemFunc[]) {
+  constructor(...systems: SystemFunc[]) {
     this.systems = [];
-    this.add_systems(...systems);
-    this.last_tick_check = 0;
+    this.addSystems(...systems);
+    this.lastTickCheck = 0;
   }
 
-  public add_systems(...systems: SystemFunc[]) {
-    const new_length = systems.length + this.systems.length;
+  /**
+   * Add systems to this stage
+   * @param systems
+   */
+  addSystems(...systems: SystemFunc[]) {
+    const newLength = systems.length + this.systems.length;
 
-    if (new_length > MAX_SYSTEMS_PER_STAGE) {
+    if (newLength > MAX_SYSTEMS_PER_STAGE) {
       throw new Error(
         `Trying to add systems to stage which would exceede max system limit (${MAX_SYSTEMS_PER_STAGE})`,
       );
@@ -32,7 +44,11 @@ export class Stage implements Runnable {
     this.systems.push(...systems.map((x) => new System(x)));
   }
 
-  public run(world: World) {
+  /**
+   * Run this stage on given world
+   * @param world
+   */
+  run(world: World) {
     for (let i = 0; i < this.systems.length; i++) {
       try {
         this.systems[i].run(world);
@@ -46,26 +62,30 @@ export class Stage implements Runnable {
       }
     }
 
-    this.check_change_tick(world);
+    this.checkChangeTick(world);
   }
 
-  public check_change_tick(world: World) {
-    const change_tick = world.change_tick;
-    const ticks_since_last_check = change_tick - this.last_tick_check;
+  /**
+   * Check & validate change ticks for world
+   * @param world
+   */
+  checkChangeTick(world: World) {
+    const changeTick = world.changeTick;
+    const ticksSinceLastCheck = changeTick - this.lastTickCheck;
 
     /**
      * Only check after at least `MAX_SYSTEMS_PER_STAGE` counts, since thats the maximum
      * allowed systems in a stage and thus a safe threshold so we don't have to check every update.
      */
-    if (ticks_since_last_check > MAX_SYSTEMS_PER_STAGE) {
+    if (ticksSinceLastCheck > MAX_SYSTEMS_PER_STAGE) {
       for (let i = 0; i < this.systems.length; i++) {
-        this.systems[i].check_change_tick(change_tick);
+        this.systems[i].checkChangeTick(changeTick);
       }
 
       // also check component ticks
-      world.check_change_ticks();
+      world.checkChangeTicks();
 
-      this.last_tick_check = change_tick;
+      this.lastTickCheck = changeTick;
     }
   }
 }