+++ packages/dreki/src/scheduler/system.ts
@@ -1,7 +1,7 @@
 import { MAX_CHANGE_TICK_DELTA } from "../constants";
-import { EventsCount } from "../world/events";
+import { EventsCounter } from "../world/events";
 import { World } from "../world/mod";
-import { runtime } from "../world/runtime";
+import { Runtime } from "../world/runtime";
 import { Runnable } from "./stage";
 
 export type SystemFunc = (world: World) => unknown;
@@ -9,31 +9,35 @@ export type SystemFunc = (world: World) => unknown;
 export class System implements Runnable {
   readonly func: SystemFunc;
 
-  last_change_tick: number;
-  event_counts: EventsCount;
+  lastChangeTick: number;
+  eventsCounter: EventsCounter;
 
   constructor(func: SystemFunc) {
     this.func = func;
-    this.last_change_tick = 0;
-    this.event_counts = new Map();
+    this.lastChangeTick = 0;
+    this.eventsCounter = new Map();
   }
 
+  /**
+   * Run this system on given world
+   * @param world
+   */
   run(world: World) {
     // increment world change tick.
-    const last_change_tick = world.increment_change_tick();
+    const lastChangeTick = world.incrementChangeTick();
 
-    // set current last_change_tick runtime context to current systems tick.
-    runtime.last_change_tick = this.last_change_tick;
-    runtime.last_event_counts = this.event_counts;
+    // set current lastChangeTick Runtime context to current systems tick.
+    Runtime.lastChangeTick = this.lastChangeTick;
+    Runtime.lastEventCounts = this.eventsCounter;
 
     this.func(world);
-    this.last_change_tick = last_change_tick;
+    this.lastChangeTick = lastChangeTick;
   }
 
-  check_change_tick(change_tick: number) {
-    const tick_delta = change_tick - this.last_change_tick;
-    if (tick_delta > MAX_CHANGE_TICK_DELTA) {
-      this.last_change_tick = change_tick - MAX_CHANGE_TICK_DELTA;
+  checkChangeTick(changeTick: number) {
+    const tickDelta = changeTick - this.lastChangeTick;
+    if (tickDelta > MAX_CHANGE_TICK_DELTA) {
+      this.lastChangeTick = changeTick - MAX_CHANGE_TICK_DELTA;
     }
   }
 }