+++ packages/dreki/src/storage/components.ts
@@ -1,61 +1,60 @@
 import { Slice } from "@dreki.land/collections";
 import { Entity } from "../entity/mod";
-import {
-  ComponentInstance,
-  ComponentFlags,
-  ComponentTick,
-  ComponentInstances,
-} from "../component/mod";
+import { ComponentInstance, ComponentFlags, ComponentTick } from "../component/mod";
 import { ComponentInfo } from "../component/register";
 import { PhantomComponentStorage } from "./phantom";
 
 export type EntitySlice = Slice<Entity>;
 
+/**
+ * A storage state for a component
+ */
 export type ComponentState = readonly [
   component: ComponentInstance,
   flags: ComponentFlags,
-  added_tick: ComponentTick,
-  changed_tick: ComponentTick,
+  addedTick: ComponentTick,
+  changedTick: ComponentTick,
 ];
 
 /**
  * Generic component storage interface type.
  */
-export type ComponentStorage = {
+export interface ComponentStorage {
   insert(
     entity: Entity,
     value: ComponentInstance,
     flags: ComponentFlags,
-    change_tick: number,
+    changeTick: number,
   ): ComponentInstance;
+
   remove(entity: Entity): boolean;
   has(entity: Entity): boolean;
   get(entity: Entity): ComponentInstance;
-  get_with_state(entity: Entity): ComponentState | undefined;
+  getWithState(entity: Entity): ComponentState | undefined;
 
-  set_flag(entity: Entity, fn: (flag: ComponentFlags) => ComponentFlags): void;
-  set_added_tick(entity: Entity, changed_tick: ComponentTick): void;
-  set_changed_tick(entity: Entity, changed_tick: ComponentTick): void;
-  is_added(entity: Entity): boolean;
-  is_changed(entity: Entity): boolean;
+  setFlag(entity: Entity, fn: (flag: ComponentFlags) => ComponentFlags): void;
+  setAddedTick(entity: Entity, changedTick: ComponentTick): void;
+  setChangedTick(entity: Entity, changedTick: ComponentTick): void;
+  isAdded(entity: Entity): boolean;
+  isChanged(entity: Entity): boolean;
 
-  register_phantom(reference: PhantomComponentStorage): void;
+  registerPhantom(reference: PhantomComponentStorage): void;
 
-  check_ticks(change_tick: ComponentTick): void;
-  get_ticks(entity: Entity): readonly [added: ComponentTick, changed: ComponentTick];
+  checkTicks(changeTick: ComponentTick): void;
+  getTicks(entity: Entity): readonly [added: ComponentTick, changed: ComponentTick];
 
-  add_removed(entity: Entity, component: ComponentInstance): void;
-  has_removed(entity: Entity): boolean;
-  get_removed(entity: Entity): ComponentInstance | undefined;
-  clear_removed(): void;
+  addRemoved(entity: Entity, component: ComponentInstance): void;
+  hasRemoved(entity: Entity): boolean;
+  getRemoved(entity: Entity): ComponentInstance | undefined;
+  clearRemoved(): void;
 
   realloc?(length: number): void;
 
-  entity_slice(with_removed?: boolean): EntitySlice;
+  entitySlice(withRemoved?: boolean): EntitySlice;
 
   empty(): boolean;
   readonly length: number;
-  readonly length_with_removed: number;
+  readonly lengthWithRemoved: number;
   readonly capacity: number;
   readonly info: ComponentInfo;
-};
+}